refactor: CardSubmitPage .jsx -> .tsx
