+++ src/components/pages/CardSubmitPage/styled.tsx
